// src/types/MemberDetail.ts
export interface MemberDetail {
  requestorProfileId: string;
  requestorId: string;
  requestorProvName: string;
  requestorProvShortName: string;
  profileNumber: string;
  firstName: string;
  lastName: string;
  contactName: string;
  contactAuthority: string;
  contactEmail: string;
  contactPhoneNumber: string;
  receiverName: string;
  receiverPhoneNumber: string;
  streetNumber: string;
  streetName: string;
  city: string;
  postalCode: string;
  sgAccountId: string;
  initial: string;
}